requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
python-dotenv
google-generativeai
librosa
numpy
matplotlib
scipy
